Added download invoice feature and a filehelper function to handle the deletion of images if some image is updated or deleted

+++ controllers/admin.js
@@ -1,6 +1,8 @@
 const { ValidationErrorItem } = require('sequelize');
 const Product = require('../models/product');
 const {validationResult} = require('express-validator');
+const fileHelper = require('../util/file');
+const { file } = require('pdfkit');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -15,23 +17,39 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.file;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  console.log(imageUrl);
+  if(!image){
+    return res.render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/edit-product',
+      editing: false,
+      errorMessage: 'Invalid file format provided',
+      hasError: true,
+
+      product: {
+        title:title,
+        price:price,
+        description:description
+
+      },
+      validationError: []
+    });
+  }
+
   const errors = validationResult(req);
 
   if(!errors.isEmpty()){
-      return res.render('admin/edit-product', {
+      return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
-      path: '/admin/edit-product',
+      path: '/admin/add-product',
       editing: false,
       errorMessage: errors.array()[0].msg,
       hasError: true,
 
       product: {
         title:title,
-        imageUrl:imageUrl,
         price:price,
         description:description
 
@@ -39,6 +57,8 @@ exports.postAddProduct = (req, res, next) => {
       validationError: errors.array()
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title,
@@ -94,7 +114,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
@@ -109,7 +129,6 @@ exports.postEditProduct = (req, res, next) => {
 
       product: {
         title:updatedTitle,
-        imageUrl:updatedImageUrl,
         price:updatedPrice,
         description:updatedDesc,  
         _id:prodId
@@ -128,7 +147,12 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if(image){
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+      
+
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
@@ -164,7 +188,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id:prodId, userId: req.user._id})
+  Product.findById(prodId).then(
+    product =>{
+      if(!product){
+        return next(new Error('Product Not found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({_id:prodId, userId: req.user._id})
+   
+    })
+  
+  
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
